# Conflicts:
#	solr/core/src/java/org/apache/solr/security/RuleBasedAuthorizationPlugin.java
#	solr/solr-ref-guide/src/authentication-and-authorization-plugins.adoc
#	solr/solr-ref-guide/src/rule-based-authorization-plugin.adoc

+++ solr/core/src/test/org/apache/solr/cloud/MissingSegmentRecoveryTest.java
@@ -79,6 +79,10 @@ public class MissingSegmentRecoveryTest extends SolrCloudTestCase {
   
   @After
   public void teardown() throws Exception {
+    if (null == leader) {
+      // test did not initialize, cleanup is No-Op;
+      return;
+    }
     System.clearProperty("CoreInitFailedAction");
     CollectionAdminRequest.deleteCollection(collection).process(cluster.getSolrClient());
   }
