+++ solr/core/src/test/org/apache/solr/rest/schema/analysis/TestManagedSynonymGraphFilterFactory.java
@@ -33,8 +33,9 @@ import org.eclipse.jetty.servlet.ServletHolder;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.noggit.JSONUtil;
 import org.restlet.ext.servlet.ServerServlet;
+
+import static org.apache.solr.common.util.Utils.toJSONString;
 
 // See: https://issues.apache.org/jira/browse/SOLR-12028 Tests cannot remove files on Windows machines occasionally
 public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
@@ -62,9 +63,13 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
 
   @After
   private void after() throws Exception {
-    jetty.stop();
-    jetty = null;
-    FileUtils.deleteDirectory(tmpSolrHome);
+    if (null != jetty) {
+      jetty.stop();
+      jetty = null;
+    }
+    if (null != tmpSolrHome) {
+      FileUtils.deleteDirectory(tmpSolrHome);
+    }
     System.clearProperty("managed.schema.mutable");
     System.clearProperty("enable.update.log");
 
@@ -91,7 +96,7 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     Map<String,List<String>> syns = new HashMap<>();
     syns.put("happy", Arrays.asList("glad","cheerful","joyful"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     assertJQ(endpoint,
@@ -117,7 +122,7 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     syns.put("sad", Arrays.asList("unhappy"));
     syns.put("SAD", Arrays.asList("bummed"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     assertJQ(endpoint,
@@ -187,7 +192,7 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     syns = new HashMap<>();
     syns.put("mad", Arrays.asList("angry"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     assertJQ(endpoint,
@@ -197,7 +202,7 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     syns = new HashMap<>();
     syns.put(multiTermSynonym, Arrays.asList(multiTermOrigin));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     assertJQ(endpoint+"/"+URLEncoder.encode(multiTermSynonym, "UTF-8"),
@@ -233,12 +238,12 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     syns = new HashMap<>();
     syns.put("mb", Arrays.asList("megabyte"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     syns.put("MB", Arrays.asList("MiB", "Megabyte"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     assertJQ(endpoint + "/MB",
@@ -247,7 +252,7 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     // test for SOLR-6878 - by default, expand is true, but only applies when sending in a list
     List<String> m2mSyns = new ArrayList<>();
     m2mSyns.addAll(Arrays.asList("funny", "entertaining", "whimiscal", "jocular"));
-    assertJPut(endpoint, JSONUtil.toJSON(m2mSyns), "/responseHeader/status==0");
+    assertJPut(endpoint, toJSONString(m2mSyns), "/responseHeader/status==0");
 
     assertJQ(endpoint + "/funny",
         "/funny==['entertaining','funny','jocular','whimiscal']");
@@ -279,7 +284,7 @@ public class TestManagedSynonymGraphFilterFactory extends RestTestBase {
     // now put a synonym
     syns.put("fröhlich", Arrays.asList("glücklick"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     // and check if it exists
