+++ solr/core/src/test/org/apache/solr/util/UtilsToolTest.java
@@ -77,7 +77,10 @@ public class UtilsToolTest extends SolrTestCaseJ4 {
   @After
   public void tearDown() throws Exception {
     super.tearDown();
-    org.apache.commons.io.FileUtils.deleteDirectory(dir.toFile());
+    if (null != dir) {
+      org.apache.commons.io.FileUtils.deleteDirectory(dir.toFile());
+      dir = null;
+    }
   }
   
   @Test
